requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.2.1
pydantic[email]==2.9.2
email-validator==2.1.0.post1
sqlalchemy==2.0.35
pymysql==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
Faker==24.4.0